Changed lighting, added 5 new avatars, added random wait time to script
