+++ Assets/Code/SupermarketLevel/Cashier.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f265e60dfb08a0543ac4fe2aa48dbad0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
