+++ Assets/Code/SupermarketLevel/GoalPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8c1e746f5e4cac4ea01e50a6b5b0858
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
