--- Assets/animationStateController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c635e0cc7e353644fb581730b508f673
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
